harden: i64 align scores, exact-LLL dim cap, align memory bounds
Hardening pass over the recently-built lattice/alignment/rational surface. No
behavior change for valid inputs — closes silent-overflow and resource-exhaustion
edges that custom align scoring and exact LLL opened.

- align scores widened i32 -> i64 (src/align.rs: NEG_INF, Scoring, Alignment.score,
  max3, align_path). Custom weights are capped at ±10^6 and the alignment path can be
  tens of millions of cells long, so an i32 accumulator could overflow and silently
  wrap the score negative. i64 has ample headroom (worst case ~10^14 vs 2.3·10^18).
  Dropped the now-redundant `as i64` casts at the two result sites.
- align DP-table cell cap halved 100M -> 50M (both the general builtin and the DNA
  method): i64 scores double the per-cell memory (~27 B/cell across six matrices), so
  50M keeps the worst-case table near ~1.3 GB instead of ~2.7 GB.
- lll_exact gets its own dimension cap (256) instead of sharing the f64 path's 1024:
  exact Gram–Schmidt numerators grow with dimension (Hadamard bound), so a high-dim
  integer lattice could blow up memory/time far worse than the float path. Integer-
  relation detection needs only a handful of dimensions, so 256 is generous.

Already-solid (verified, not changed): rational `**` caps |exp| at 4096 (no BigInt
blow-up), rational `/` and `//` guard a zero divisor, `0 ** negative` errors,
lll_exact rejects fractional entries and i64-overflowing output.

Test: hardening_overflow_and_resource_caps (interp) — proves a 2200×10^6 = 2.2e9 score
(would wrap in i32) is exact; exponent cap, rational /0, weight cap, cell cap, and
fractional-entry rejection all error cleanly. 242 bin + 74 cli, clippy 0, vmparity
RESULT=0.

+++ src/align.rs
@@ -11,7 +11,9 @@
 //! `O(n·m)`; fine for reads-vs-genes, not whole chromosomes (ADR 0015).
 
 /// A sentinel "unreachable" score with headroom so repeated additions never wrap.
-const NEG_INF: i32 = i32::MIN / 4;
+/// `i64` (not `i32`): with custom weights up to ±10⁶ and an alignment path up to the
+/// cell cap, an `i32` accumulator could overflow and silently corrupt the score.
+const NEG_INF: i64 = i64::MIN / 4;
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub enum Mode {
@@ -27,10 +29,10 @@ pub enum Mode {
 /// length `L` costs `gap_open + L * gap_extend`.
 #[derive(Clone, Copy)]
 pub struct Scoring {
-    pub match_: i32,
-    pub mismatch: i32,
-    pub gap_open: i32,
-    pub gap_extend: i32,
+    pub match_: i64,
+    pub mismatch: i64,
+    pub gap_open: i64,
+    pub gap_extend: i64,
 }
 
 impl Scoring {
@@ -45,7 +47,7 @@ impl Scoring {
 /// (`=`/`X`/`I`/`D`, the same op characters the SAM reader renders), the two
 /// gap-padded sequences, and the `[start, end)` span covered in the target.
 pub struct Alignment {
-    pub score: i32,
+    pub score: i64,
     pub cigar: String,
     pub x_aligned: String,
     pub y_aligned: String,
@@ -55,7 +57,7 @@ pub struct Alignment {
 
 /// Pick the maximum of three scores, returning it and which state won (0=M,1=X,2=Y).
 /// Ties prefer M, then X — keeping diagonal (aligned) moves over gaps.
-fn max3(a: i32, b: i32, c: i32) -> (i32, u8) {
+fn max3(a: i64, b: i64, c: i64) -> (i64, u8) {
     if a >= b && a >= c {
         (a, 0)
     } else if b >= c {
@@ -159,7 +161,7 @@ pub fn align(x: &[u8], y: &[u8], mode: Mode, sc: Scoring) -> Alignment {
         }
         Mode::Local => {
             // The highest M cell anywhere.
-            let (mut bi, mut bj, mut best) = (0usize, 0usize, 0i32);
+            let (mut bi, mut bj, mut best) = (0usize, 0usize, 0i64);
             for ii in 0..=n {
                 for jj in 0..=m {
                     if mm[ii * w + jj] > best {
@@ -292,7 +294,7 @@ pub fn align_path(
     mode: Mode,
     sc: Scoring,
     eq: impl Fn(usize, usize) -> bool,
-) -> (i32, Vec<Step>) {
+) -> (i64, Vec<Step>) {
     let w = m + 1;
     let size = (n + 1) * w;
     let mut mm = vec![NEG_INF; size];
@@ -367,7 +369,7 @@ pub fn align_path(
             (n, m, st, sc2)
         }
         Mode::Local => {
-            let (mut bi, mut bj, mut best) = (0usize, 0usize, 0i32);
+            let (mut bi, mut bj, mut best) = (0usize, 0usize, 0i64);
             for ii in 0..=n {
                 for jj in 0..=m {
                     if mm[ii * w + jj] > best {
